+++ spec/spec_helper.rb
@@ -15,6 +15,7 @@ RSpec.configure do |config|
 end
 
 ActiveJob::Base.queue_adapter = :sidekiq
+ActiveJob::Base.logger.level = :warn
 
 if Gem::Dependency.new('sidekiq', '>= 5.0.0').matching_specs.any?
   require 'active_record'